refactor: move pool-add components
