--- packages/web/src/components/pool/tooltip-content/TooltipContentDummy.ts
@@ -1,236 +1,0 @@
-import { TITLE } from "./TooltipContent";
-
-export const COMPOSITION = {
-  type: TITLE.COMPOSITION,
-  content: [
-    {
-      title: "Composition",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          rate: "50%",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          rate: "50%",
-          amount: "150.0255",
-        },
-      ],
-    },
-  ],
-};
-
-export const CLAIM_REWARDS = {
-  type: TITLE.CLAIM_REWARDS,
-  content: [
-    {
-      title: "Swap Fees",
-      total: "$150.21",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-    {
-      title: "Staking rewards",
-      total: "$150.21",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-    {
-      title: "Unclaimable",
-      total: "$150.21",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-  ],
-};
-
-export const BALANCE = {
-  type: TITLE.BALANCE,
-  content: [
-    {
-      title: "Balance",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-  ],
-};
-
-export const TOTAL_REWARDS = {
-  type: TITLE.TOTAL_REWARDS,
-  content: [
-    {
-      title: "Swap Fees",
-      total: "$150.21",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-    {
-      title: "Staking rewards",
-      total: "$150.21",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-  ],
-};
-
-export const ESTIMATED_APR = {
-  type: TITLE.ESTIMATED_APR,
-  content: [
-    {
-      title: "APR Breakdown",
-      swapFees: "50.05%",
-      stakingRewards: "50.05%",
-    },
-  ],
-};
-
-export const BEING_UNSTAKED = {
-  type: TITLE.BEING_UNSTAKED,
-  content: [
-    {
-      title: "Available in 20 days",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-    {
-      title: "Available in 16 days",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-    {
-      title: "Available in 23 hours",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-    {
-      title: "Available in 58 minutes",
-      tokens: [
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-          symbol: "HEX",
-          amount: "50.05881",
-        },
-        {
-          logoURI:
-            "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-          symbol: "USDC",
-          amount: "150.0255",
-        },
-      ],
-    },
-  ],
-};
