--- packages/web/src/containers/add-incentive-banner-container/AddIncentiveBannerContainer.tsx
@@ -1,8 +1,0 @@
-import AddIncentiveBanner from "@components/pool/add-incentive-banner/AddIncentiveBanner";
-import React from "react";
-
-const AddIncentiveBannerContainer: React.FC = () => {
-  return <AddIncentiveBanner />;
-};
-
-export default AddIncentiveBannerContainer;
