+++ packages/web/src/views/pool/pool-add/components/exchange-rate-graph/ExchangeRateGraph.styles.ts
