+++ packages/web/src/views/pool/pool-add/components/exchange-rate-graph/exchange-rate-graph-content/ExchangeRateGraphContent.styles.ts
